task-4-dashboard/pyproject.toml
[tool.poetry]
name = "task-4-dashboard"
version = "0.1.0"
description = ""
authors = ["exkuretrol <33695301+exkuretrol@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
gunicorn = "^21.2.0"
environs = {extras = ["django"], version = "^10.3.0"}
psycopg = {extras = ["binary"], version = "^3.1.17"}
whitenoise = "^6.6.0"
django-widget-tweaks = "^1.5.0"
pillow = "^10.2.0"
django-browser-reload = "^1.12.1"
django-seed = "^0.3.1"
djangorestframework = "^3.14.0"
django-tables2 = "^2.7.0"
django-filter = "^23.5"
django-debug-toolbar = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
